requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
redis==4.5.4
mlflow==2.3.1
scikit-learn==1.0.2
numpy==1.26.4
pandas==1.3.3
torch==1.9.0
faiss-cpu==1.7.4
prometheus-client==0.11.0
python-dotenv==0.19.0
requests==2.31.0